Add Client logic to call the server
