+++ kotlin/thrift-example/client/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    application
+}
+
+application {
+    mainClassName = "thriftclient.Client"
+}
+
+dependencies {
+    implementation(project(":model"))
+}
