+++ kotlin/thrift-example/server/src/main/kotlin/thriftexample/FortuneTellerService.kt
@@ -37,4 +37,3 @@ class FortuneTellerService : FortuneTeller.Iface {
         }
     }
 }
-
