+++ kotlin/thrift-example/settings.gradle.kts
@@ -2,5 +2,6 @@ rootProject.name = "thrift-example"
 
 include(
     ":model",
-    ":server"
+    ":server",
+    ":client"
 )
